Fixed some build warnings. Enabled Ogre viewer by default, again.

+++ trunk/core/smartbody/sbgui/src/fltk_viewer.cpp
@@ -2282,7 +2282,7 @@ int FltkViewer::handle ( int event )
 					if (box.volume() < .0001)
 					{
 						double val = 1.0 / SmartBody::SBScene::getScene()->getScale() * .5;
-						box.grows(val, val, val);
+						box.grows((float) val, (float) val, (float) val);
 					} 
 					sceneBox.extend(box);
 				}
